Fix misaligned load on Sparc CPU causing segfault
Transform all uses of {be,le}::swap of the forms:
  swap(*(T *)(pointer expr)) -> peek<T>(pointer expr)
  swap(<array of T>[i])      -> peek<T>(<array of T> + i)
The calculated address derefernced before calling swap may be missaligned and in these cases it is possible for compiler to not recognise this (especially where the address has been cast from void * or byte *) and emit a misaligned load. As peek (and by extension read) access memory as bytes the are immune from alignment issues.
All other uses of swap are (e.g. on packed structure members) is fine as the compiler can recognise a misaligned load and emit the correct instructions to compensate.

+++ src/inc/Pass.h
@@ -56,13 +56,13 @@ private:
     bool   	testPassConstraint(vm::Machine & m) const;
     bool   	testConstraint(const Rule & r, vm::Machine &) const;
     bool   	readFSM(const byte* p, const byte*const pass_start, const size_t max_offset);
-    bool   	readRules(const uint16 * rule_map, const size_t num_entries,
+    bool   	readRules(const byte * rule_map, const size_t num_entries,
                      const byte *precontext, const uint16 * sort_key,
                      const uint16 * o_constraint, const byte *constraint_data, 
                      const uint16 * o_action, const byte * action_data,
                      const Face &);
-    bool   	readStates(const int16 * starts, const int16 * states, const uint16 * o_rule_map);
-    bool   	readRanges(const uint16* ranges, size_t num_ranges);
+    bool   	readStates(const byte * starts, const byte * states, const byte * o_rule_map);
+    bool   	readRanges(const byte * ranges, size_t num_ranges);
     uint16 	glyphToCol(const uint16 gid) const;
     bool   	runFSM(FiniteStateMachine & fsm, Slot * slot) const;
     void	dumpRuleEventConsidered(const FiniteStateMachine & fsm, const RuleEntry & re) const;
